+++ benchmarks/llm-sort-floating/gemini-sort.h
@@ -3,16 +3,16 @@
 
 class Gemini {
   public:
-    static void bubbleSort(double arr[], int n);
-    static void insertionSort(double arr[], int n);
-    static void mergeSort(double arr[], int n);
-    static void quickSort(double arr[], int n);
-    static void heapSort(double arr[], int n);
-    static void gnomeSort(double arr[], int n);
-    static void radixSort(double arr[], int n);
-    static void shellSort(double arr[], int n);
-    static void combSort(double arr[], int n);
-    static void pancakeSort(double arr[], int n);
+    static void bubbleSort(float arr[], int n);
+    static void insertionSort(float arr[], int n);
+    static void mergeSort(float arr[], int n);
+    static void quickSort(float arr[], int n);
+    static void heapSort(float arr[], int n);
+    static void gnomeSort(float arr[], int n);
+    static void radixSort(float arr[], int n);
+    static void shellSort(float arr[], int n);
+    static void combSort(float arr[], int n);
+    static void pancakeSort(float arr[], int n);
 };
 
 #endif
